backend/pyproject.toml
[tool.poetry]
name = "learnmate-backend"
version = "0.1.0"
description = "Backend for LearnMate AI"
authors = ["Jaime Mulero <jmulero@lunia.es>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.3"
python-dotenv = "1.0.0"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
openai = "^1.52.2"
python-multipart = "0.0.12"
apscheduler = "^3.10.4"
langchain-community = "^0.3.3"
langchain-openai = "^0.2.3"
langchain-text-splitters = "^0.3.0"
faiss-cpu = "^1.9.0"
llama-parse = "^0.5.12"
unstructured = "^0.16.3"
markdown = "^3.7"
unstructured-client = "^0.26.1"
azure-ai-documentintelligence = "^1.0.0b4"
pdfminer-six = "^20240706"
pi-heif = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
